Refactored the PCQBackend project to remove demo reference and renamed package
