+++ bin/init.sh
@@ -25,7 +25,7 @@ then
 fi
 
 declare -a files_with_port=(.env Dockerfile README.md src/main/resources/application.yaml)
-declare -a files_with_slug=(build.gradle docker-compose.yml Dockerfile README.md web.config ./infrastructure/main.tf ./src/main/java/uk/gov/hmcts/reform/demo/controllers/RootController.java)
+declare -a files_with_slug=(build.gradle docker-compose.yml Dockerfile README.md web.config ./infrastructure/main.tf ./src/main/java/uk/gov/hmcts/reform/pcqbackend/controllers/RootController.java)
 
 # Replace port number
 for i in ${files_with_port[@]}
