+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
   id 'pmd'
   id 'jacoco'
   id 'io.spring.dependency-management' version '1.0.9.RELEASE'
-  id 'org.springframework.boot' version '2.2.4.RELEASE'
+  id 'org.springframework.boot' version '2.2.5.RELEASE'
   id 'org.owasp.dependencycheck' version '5.3.0'
   id 'com.github.ben-manes.versions' version '0.28.0'
   id 'org.sonarqube' version '2.8'
@@ -19,6 +19,16 @@ allprojects {
 }
 
 sourceSets {
+
+  unitTest {
+    java {
+      compileClasspath += main.output
+      runtimeClasspath += main.output
+      srcDir file('src/test/java')
+    }
+    resources.srcDir file('src/test/resources')
+  }
+
   functionalTest {
     java {
       compileClasspath += main.output
@@ -49,6 +59,11 @@ sourceSets {
 
 tasks.withType(JavaCompile) {
   options.compilerArgs << "-Xlint:unchecked" << "-Werror"
+}
+
+task unit(type: Test, description: 'Runs the unit tests.', group: 'Verification') {
+  testClassesDirs = sourceSets.test.output.classesDirs
+  classpath = sourceSets.test.runtimeClasspath
 }
 
 task functional(type: Test, description: 'Runs the functional tests.', group: 'Verification') {
@@ -186,7 +201,7 @@ test {
   failFast = true
 }
 
-mainClassName = 'uk.gov.hmcts.reform.demo.Application'
+mainClassName = 'uk.gov.hmcts.reform.pcqbackend.Application'
 
 bootJar {
   archiveName = 'pcq-backend.jar'
